requirements.txt

streamlit 
torch 
torchvision 
pillow 
streamlit-drawable-canvas
streamlit 
